Re-position y-axis of each channel, align to grid
This is a major overhaul of aligning individual channels
to the visible y-axis grid (at least with default gain).
This fixes various rounding-errors in grid-spacing
particularly when the window is resized.

The y-axis dial now refers to the absolute position,
this breaks the state of existing instances.
This is only for visualization, so deemed acceptable with
a minor version update.